Renamed the header consistently with the last project additions
